add tablets and accessories page
