+++ src/components/Catalog/Catalog.module.scss
@@ -1,0 +1,137 @@
+.phones {
+  @include container;
+
+  padding-top: 24px;
+  &__routs {
+    display: flex;
+    align-items: center;
+    &__home {
+      width: 16px;
+      height: 16px;
+      background-image: url(../../img/icon/Home.png);
+      background-size: cover;
+      background-repeat: no-repeat;
+      background-position: center;
+    }
+    &__array {
+      box-sizing: border-box;
+      padding: 0 8px;
+      width: 16px;
+      height: 16px;
+      background-image: url(../../img/arrow/ArrowRight.png);
+      background-size: cover;
+      background-repeat: no-repeat;
+      background-position: center;
+    }
+    &__currentPage {
+      font-size: 12px;
+      font-weight: 600;
+      line-height: 15px;
+      color: $secondary-color;
+    }
+  }
+  &__title {
+    margin-top: 24px;
+    margin-bottom: 8px;
+    font-weight: 800;
+    font-size: 32px;
+    line-height: 41px;
+    letter-spacing: -1%;
+    color: $primary-color;
+    @include onTablet {
+      margin-top: 40px;
+      font-size: 48px;
+      line-height: 56px;
+    }
+  }
+  &__countModel {
+    font-weight: 600;
+    font-size: 14px;
+    color: $secondary-color;
+  }
+  &__items {
+    list-style: none;
+    @include mainPageGrid;
+    row-gap: 40px;
+    margin-top: 24px;
+    align-items: stretch;
+  }
+  &__item {
+    grid-column: span 4;
+    height: auto;
+
+    @include onTablet {
+      grid-column: span 4;
+    }
+    @include onDesktop {
+      grid-column: span 6;
+    }
+  }
+  &__pagination {
+    display: flex;
+
+    justify-content: center;
+    margin-top: 40px;
+    column-gap: 16px;
+
+    &__arr {
+      width: 36px;
+      height: 36px;
+      background-color: #fff;
+      border: 1px solid $icons-color;
+      border-radius: 48px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      &:hover {
+        border: 1px solid $primary-color;
+      }
+
+      &Left {
+        width: 16px;
+        height: 16px;
+        background-image: url(../../img/arrowActive/ArrowLeft.png);
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-position: center;
+      }
+      &Right {
+        width: 16px;
+        height: 16px;
+        background-image: url(../../img/arrowActive/ArrowRight.png);
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-position: center;
+      }
+    }
+    &__buttons {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      flex-wrap: wrap;
+      column-gap: 8px;
+    }
+
+    &__button {
+      background-color: #fff;
+      width: 32px;
+      height: 32px;
+      border-radius: 16px;
+      border: 1px solid $elemnts-color;
+      color: $primary-color;
+      font-size: 14px;
+      font-weight: 600;
+      line-height: 21px;
+      &:hover {
+        border: 1px solid $primary-color;
+      }
+      &_current {
+        width: 32px;
+        height: 32px;
+        border-radius: 16px;
+        color: #fff;
+        background-color: $primary-color;
+      }
+    }
+  }
+}
