+++ src/components/Main/CategoriesList/CategoriesList.scss
@@ -1,6 +1,8 @@
 .categories {
   @include container;
   &__title {
+    margin-top: 40px;
+    margin-bottom: 24px;
   }
   &__list {
     margin: 0;
