+++ src/components/Main/Main.scss
@@ -149,16 +149,18 @@
   transition: 1s;
   box-sizing: border-box;
   @include cardWidth;
+  height: 100%;
 
   &:hover {
     transform: scale(105%);
   }
 
   &__logo {
-    max-height: auto;
-    width: 100%;
+    height: 196px;
+    width: auto;
     flex-shrink: 1;
     align-self: center;
+    margin-bottom: auto;
   }
 
   &__title {
@@ -170,7 +172,7 @@
   }
 
   &__price {
-    margin-top: 0;
+    margin-top: auto;
     font-size: 22px;
     font-weight: 800;
     color: #0f0f11;
@@ -186,7 +188,7 @@
     flex-direction: column;
     list-style: none;
     justify-content: center;
-    margin: 0;
+    margin-top: auto;
     padding: 0;
   }
 
@@ -211,7 +213,7 @@
   }
 
   &__interactive-part {
-    margin-top: 8px;
+    margin-top: auto;
     display: flex;
     justify-content: space-between;
 
@@ -219,5 +221,8 @@
       margin: 0;
       padding: 0;
     }
+  }
+  &__inform {
+    margin-top: auto;
   }
 }
