+++ src/pages/AccessoriesPage/AccessoriesPage.tsx
@@ -1,7 +1,23 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useAppDispatch, useAppSelector } from '../../app/reduxHooks';
+import { RootState } from '../../app/store';
+import { loadAccessories } from '../../app/slices/accessoriesSlice';
+import { Catalog } from '../../components/Catalog/Catalog';
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
 type Props = {};
 
 export const AccessoriesPage: React.FC<Props> = () => {
-  return <>Accessories page</>;
+  const dispatch = useAppDispatch();
+  const accessories = useAppSelector(
+    (state: RootState) => state.accessories.accessories,
+  );
+
+  useEffect(() => {
+    dispatch(loadAccessories());
+  }, [dispatch]);
+  return (
+    <>
+      <Catalog items={accessories} />
+    </>
+  );
 };
