+++ src/pages/TabletspPage/TabletsPage.tsx
@@ -1,7 +1,21 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { Catalog } from '../../components/Catalog/Catalog';
+import { useAppDispatch, useAppSelector } from '../../app/reduxHooks';
+import { RootState } from '../../app/store';
+import { loadTablets } from '../../app/slices/tabletsSlice';
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
 type Props = {};
 
 export const TabletsPage: React.FC<Props> = () => {
-  return <>Tablets page</>;
+  const dispatch = useAppDispatch();
+  const tablets = useAppSelector((state: RootState) => state.tablets.tablets);
+
+  useEffect(() => {
+    dispatch(loadTablets());
+  }, [dispatch]);
+  return (
+    <>
+      <Catalog items={tablets} />
+    </>
+  );
 };
