+++ src/styles/utils/_mixins.scss
@@ -49,7 +49,7 @@
 
 @mixin cardItemSize() {
   width: 100%;
-  height: auto;
+  height: 100%;
   @include onMobile {
     // width: 212px;
     // height: 439px;
